Create QuizAdapter to remove the delegation responsibility from QuizComposer.
